Def

Documentation

data Tile

A tile is a 3D object with a texture, a set of rules and a character representation.

Constructors

Tile

Fields

materials :: Map Side Material

The FilePath to a texture location, which is used in the Output.

rules :: Rule

The Rule that determines if a tile can be placed. Rules can be composed using the functions from the CompareRule type class.

charRep :: Char

The character representation is only used for debugging purposes.

data Material

Constructors

Material

Fields

ambientColor :: (Float, Float, Float)

diffuseColor :: (Float, Float, Float)

specularColor :: (Float, Float, Float)

transparency :: Float

specularExponent :: Float

illuminationModel :: Int

texture :: Maybe FilePath

extraFields :: [String]

extraFiles :: [FilePath]

data Side

Constructors

PosX

NegX

PosY

NegY

PosZ

NegZ

newtype Rule

A rule is a function that takes a TileMap and a position and returns a RuleResult.

Constructors

Rule (TileMap -> Pos -> RuleResult)

data RuleResult

The result of a rule is defined as a RuleResult. It is represented either as a `CanPlace Bool` which means the rule guaranteed passes or fails. Or as `ChancePlace Float` which gives a chance between 0 and 1 that the rule passes (1 being 100%).

Constructors

CanPlace Bool
CanPlace simply specifies if a tile can be placed at a position or not. When it is true, it has weight 1, when it is false it has weight 0

ChancePlace Float
ChancePlace specifies the chance that a tile can be placed at a position. If it is 0, it will never be placed. Otherwise, it represents the weight that is used in the generator.

resultToBool :: RuleResult -> Bool

Converts a RuleResult to a boolean, which is true if the result is true or the chance that it is placed is greater than 0.

resultToFloat :: RuleResult -> Float

Convrts a RuleResult to a float, which returns the chance if it is a chance or when it is a boolean it returns 1 for true and 0 for false.

class CompareRule a where

Methods

(<||>) :: a -> a -> a

The OR operator for rules

(<&&>) :: a -> a -> a

The AND operator for rules

(<!>) :: a -> a

The NOT operator for rules

type Pos = (Int, Int, Int)

A position is a 3D coordinate in the world

type Size = ((Int, Int, Int), (Int, Int, Int))

A size is a 3D coordinate representing the minimum and maximum coordinates of the world

newtype World

A world is a 3D grid of tiles, represented as a size of the world and a tilemap containing the tiles at each position in the world

Constructors

World (Size, TileMap)

newtype TileMap

A tilemap is a map of positions to tiles in the world

Constructors

TileMap (Map Pos Tile)

type Shape = Pos -> [Pos]

A shape is a function that takes a position and returns a list of absolute positions that are relative to the given position, forming a shape. allNeighbours is an example of a shape.