Examples.NoGreen

Description

This world becomes impossible whenever a green tile is placed. This is a useful example to benchmark very difficult to create worlds and test backtracking.

Documentation

redMaterial :: Material

redTile :: Tile

A tile that is red and can not be placed next to a green tile

blueMaterial :: Material

blueTile :: Tile

A tile that is blue and can not be placed next to a green tile

greenMaterial :: Material

greenTile :: Tile

A tile that is green and can not be placed next to a green tile

allTiles :: [Tile]

The tiles that will be used in the wave function collapse algorithm to generate the world