Examples.Roads

Description

A module that defines a set of tiles that can be used to generate a road-like structure. The roads must be connected to each other, and the road can have a turn, a split, a cross, a straight line, or an elbow. The world can also contain dirt, grass and water.

Documentation

dirtMaterial :: Material

grassDirtMaterialSide :: Material

topLevelBlock :: Map Side Material

dirtBlock :: Map Side Material

waterMaterial :: Material

waterBlock :: Map Side Material

posTopRule :: Rule

verticalNeighbour :: Shape

horizontalNeighbour :: Shape

leftConnection :: [Tile]

rightConnection :: [Tile]

upConnection :: [Tile]

downConnection :: [Tile]

upConnectionRule :: Rule

downConnectionRule :: Rule

rightConnectionRule :: Rule

leftConnectionRule :: Rule

noUpConnectionRule :: Rule

noDownConnectionRule :: Rule

noRightConnectionRule :: Rule

noLeftConnectionRule :: Rule

topCrossTile :: Tile

topBLUTSplitTile :: Tile

Top left up top split

topURBTSplitTile :: Tile

topLURTSplitTile :: Tile

topLRBTSplitTile :: Tile

topLDElbowTile :: Tile

topDRElbowTile :: Tile

topULElbowTile :: Tile

topURElbowTile :: Tile

topHorizontalPipeTile :: Tile

topVerticalPipeTile :: Tile

topGrassTile :: Tile

dirtTile :: Tile

waterTile :: Tile

allTiles :: [Tile]