Gen

Documentation

waveFuncCollapse :: [Tile] -> Size -> IO TileMap

The wave function collapse algorithm is a way to generate a tilemap based on a set of rules. The algorithm works by starting with a completely empty tilemap and then collapsing the wave function of each tile based on the rules and the surrounding tiles. This is done until the entire tilemap is filled.

takeAndRemove :: Int -> [a] -> (a, [a])

waveFuncCollapse' :: [Pos] -> TileMap -> Map Pos [Tile] -> IO TileMap

superPosition :: [Tile] -> Size -> Map Pos [Tile]

waveFuncCollapseStep :: Pos -> TileMap -> Map Pos [Tile] -> IO (Maybe (TileMap, Map Pos [Tile]))

randomTile :: Pos -> [Tile] -> TileMap -> IO (Maybe Tile)

weightToTile :: [Tile] -> [Float] -> Float -> Maybe Tile

generate :: [Tile] -> TileMap

isHigherThanRandom :: Float -> Bool