Gen.PerlinNoise

Description

This is a generator that generates a heightmap using Perlin noise. The HeightMap can be used to generate a world using the Gen.WaveFuncCollapse module, or it can be used to generate an image using the Output module.

Documentation

data PerlinConfig

The configuration for the Perlin noise generator. The seed is the seed used to generate the permutations. The seed can also be generated randomly using the perlinNoiseRandom function.

Constructors

PerlinConfig

Fields

seed :: Int

permSize :: Int

octaves :: Int

scale :: Float

persistence :: Float

frequency :: Float

perlinNoise :: PerlinConfig -> HeightMap

Generate a heightmap using Perlin noise.

perlinNoiseRandom :: PerlinConfig -> IO HeightMap

Generate a heightmap using Perlin noise with a random seed. | The seed value in the PerlinConfig is replaced with a random seed.