Gen.WaveFuncCollapse

Description

The wave function collapse algorithm is a way to generate a tilemap based on a set of rules. The algorithm works by starting with a completely empty tilemap and then collapsing the wave function of each tile based on the rules and the surrounding tiles. This is done until the entire tilemap is filled.

Documentation

type Weights = (Float, [Float])

Weights is a tuple of a float and a list of floats (Total weight, [Weights])

type ShannonEntropy = Float

type Env = Map Pos ([Tile], Weights, ShannonEntropy)

waveFuncCollapse :: [Tile] -> Size -> IO TileMap

Initialize the wave function collapse algorithm with a list of tiles and a size for the world. The algorithm will then generate a tilemap and an environemnt using the createEnv function. If no Environment can be created with the current Tile set, the function throw an error. Afterwards it will invoke waveFuncCollapse' which will iteratiively collapse the wave function.

createEnv :: [Tile] -> TileMap -> [Pos] -> Maybe (TileMap, Env)

Create the tileMap and environment for the wave function collapse algorithm It applies the rule to each position and creates the environment based on the result. If an environment only has one possible tile it will be added to the tileMap, and removed from the environment. If an environment has no possible tiles, the function will return Nothing.

shannonEntropy :: Weights -> ShannonEntropy

Calculate the shannon entropy of a set of tiles. We use type Weights instead of a list of tiles to avoid reevaluating the rules of the tiles and recalculating the total weight of these tiles.

waveFuncCollapse' :: [Tile] -> TileMap -> Env -> IO TileMap

The helper function of the waveFuncCollape function. It will iteratively call the waveFuncCollapseStep function until there are no more available positions in the environment. This function is called with the position where the wave function should collapse. This position is generated using the shannonPos function. If the algorithm gets stuck, it will call the resetWaveFuncCollapse function.

resetWaveFuncCollapse :: IO TileMap

Reset the wave function collapse algorithm. This is done when the algorithm gets stuck and can't continue. Not implemented yet. Optimally the algorithm should be able to backtrack to a previously solvable state.

shannonPos :: Env -> IO Pos

Get the position with the lowest shannon entropy. If there are multiple positions with the same entropy, the function will return a random position from the list of positions.

getRandomElement :: [a] -> IO a

Get a random element from a list

waveFuncCollapseStep :: [Tile] -> Pos -> TileMap -> Env -> IO (Maybe (TileMap, Env))

Collapse the wave function of a single position. Selects a random tile from the environment and adds it to the tileMap. If the environment has no possible tiles, the function will return Nothing. After the tile has been added to the tileMap, the position will be removed from the environment, and the environment will be updated with the new possible tiles.

randomTile :: [Tile] -> Weights -> IO (Maybe Tile)

Select a random tile from a list of tiles based on their weights. If the total weight is 0, the function will return Nothing.

weightToTile :: [Tile] -> [Float] -> Float -> Maybe Tile

Convert a list of tiles and their weights to a single tile based on a random number. The function will return Nothing if list of tiles and weights are empty or the random number is greater than the total weight. The length of the tiles and weights must be the same.