Output

Description

This module contains functions to convert a world to an obj file, which can be used to render the world in a 3D renderer. The obj file can be saved, without materials, to a file using the saveWorldToObj function, or with materials to a folder using the saveWorldToObjAndMtl function.

Documentation

saveWorldToObj

Arguments

:: World

-> FilePath
The FilePath to save the obj file to

-> Float
The scale of the faces of the tiles

-> IO ()

Converts a world to an obj file, which can be used to render the world in a 3D renderer. The obj file is saved to the given path

saveWorldToObjAndMtl

Arguments

:: World

-> FilePath
The FilePath that refers to the folder where the obj and mtl files are saved, as well as the textures folder

-> Float
The scale of the faces of the tiles, which is important for the texture scaling

-> Bool
Whether to copy the textures to the textures folder, or refer to the original location

-> IO ()

Converts a world to an obj file and a mtl file, which can be used to render the world in a 3D renderer. The obj file and mtl file are saved to the given path to a folder, and the textures are copied to a textures folder in the given path

worldToObj :: World -> Float -> String

Converts a world to an obj file as a string, without any materials

worldToObjAndMtl :: World -> Float -> (String, String, [FilePath])

Converts a world to a tuple of strings, where the first string is the obj file, the second string is the mtl file and the third is a list of file paths to the textures used in the mtl file