Utils

Description

This module contains utility functions and data structures that can be used to define rules in an easier way. Also contains functions that can be used to create shapes that can be used in the rules, and a function that can be used to create a simple material map for all sides.

Documentation

listToShape :: [Pos] -> Shape

Convert a list of relative positions into a shape that takes a position and returns the absolute positions of the shape at that position

allNeighbours :: Shape

A shape that contains all the direct horizontal, vertical and diagonal neighbours of a position

directNeighbours :: Shape

leftNeighbour :: Shape

A shape that contains the neighbour to the left of a position

rightNeighbour :: Shape

A shape that contains the neighbour to the right of a position

downNeighbour :: Shape

A shape that contains the neighbour to the front of a position

upNeighbour :: Shape

A shape that contains the neighbour to the back of a position

aboveNeighbour :: Shape

A shape that contains the neighbour above a position

belowNeighbour :: Shape

A shape that contains the neighbour below a position

nextToAny :: [Tile] -> Shape -> Rule

A rule that takes a list of tiles and a shape, and returns True if any of the tiles defined in the list tile are within the shape at the given position.

nextToAll :: [Tile] -> Shape -> Rule

A rule that takes a list of tiles and a shape, and returns True if all of the tiles defined in the list tile are within the shape at the given position.

allMustBe :: [Tile] -> Shape -> Bool -> Rule

A rule that takes a list of tiles and a shape, and returns True if all of the tiles in the shape are within the list of tiles at the given position. The boolean parameter determines if an empty tile should be considered a match.

anyRule :: (a -> RuleMonad Bool) -> [a] -> RuleMonad Bool

allRule :: (a -> RuleMonad Bool) -> [a] -> RuleMonad Bool

weightedRule :: Float -> Rule

A rule that takes a float f and returns a rule with chance f of returning True

maxRelativeHeight :: Float -> Rule

A rule that checks if the position has a height of at most p times the height of the tileMap

minRelativeHeight :: Float -> Rule

A rule that checks if the position has a height of at least p times the height of the tileMap

isInPos :: [Pos] -> Pos -> Bool

A function that takes a list of positions and returns a function that checks if a position is in the list

canExistAt :: (Pos -> Bool) -> Rule

A rule that takes a function that takes a position and returns a boolean, and returns a rule using this function

createMaterialMapForAllSides :: Material -> Map Side Material

A function that takes a material and returns a map from all sides to that material