WorldGen.Examples.Chance

Description

This module contains examples of tiles that have a chance of being generated using the weightedRule.

Documentation

redMaterial :: Material

A solid red material

redTile :: Tile

A red tile that negates a weighted rule (! 0.90), effectively making the rule have a 10% chance of being true

blueMaterial :: Material

A solid blue material

blueTile :: Tile

A blue tile that combines two weighted rules with an <&&> operator, effectively multiplying the chances of both rules

greenMaterial :: Material

A solid green material

greenTile :: Tile

A green tile that combines two weighted rules with an <||> operator, effectively adding the chances of both rules together

airMaterial :: Material

A transparent material that represents air

airTile :: Tile

A textureless air tile that has a 30% chance of being placed

allTiles :: [Tile]

The tiles that will be used in the wave function collapse algorithm to generate the world