WorldGen.Gen.PerlinNoise

Description

This is a generator that generates a heightmap using Perlin noise. The HeightMap can be used to generate a world using the Gen.WaveFuncCollapse module, or it can be used to generate an image using the Output module.

Documentation

data PerlinConfig

The configuration for the Perlin noise generator.

Constructors

PerlinConfig

Fields

seed :: Int

The seed used to generate the permutations. Can also be generated randomly using the perlinNoiseRandom function.

permSize :: Int

The size of the permutation table. Must be greater than 0.

octaves :: Int

The number of octaves used to generate the noise. Must be greater than or equal to 0. The higher the number of octaves, the more detailed the noise, but the slower the generation.

scale :: Float

The scale of the noise. The higher the scale, the more zoomed in the noise.

persistence :: Float

The persistence of the noise. The persistence is the factor that determines how much each octave contributes to the overall shape of the noise.

frequency :: Float

The frequency of the noise. The frequency is the factor that determines how much the noise is repeated.

lacunarity :: Float

The lacunarity of the noise. The lacunarity is the factor that determines how much the frequency is increased for each octave.

perlinNoise :: PerlinConfig -> HeightMap

Generate a heightmap using Perlin noise.

perlinNoiseRandom :: PerlinConfig -> IO HeightMap

Generate a heightmap using Perlin noise with a random seed. The seed value in the PerlinConfig is replaced with a random seed.