Gen.WaveFuncCollapse

Description

The wave function collapse algorithm is a way to generate a tilemap based on a set of rules. The algorithm works by starting with a completely empty tilemap and then collapsing the wave function of each tile based on the rules and the surrounding tiles. This is done until the entire tilemap is filled.

Documentation

waveFuncCollapse

Arguments

:: [Tile]
The tiles that will be placed in the tilemap

-> Size
The size of the tilemap

-> IO (Either Error TileMap)
Returns an error if no possible tilemap can be generated, otherwise returns the tilemap

Run the wave function collapse algorithm on a list of tiles and a size for the world. The algorithm will return a tilemap that satisfies the rules of the tiles. If no possible tilemap can be generated, the function will return an error.

waveFuncCollapeHeightMap

Arguments

:: HeightMap
The heightmap that the tilemap should be generated from

-> [Tile]
The tiles that will be placed in the tilemap above the heightmap

-> [Tile]
The tiles that will be placed in the tilemap below the heightmap

-> Size
The size of the tilemap

-> IO (Either Error TileMap)
Returns an error if no possible tilemap can be generated, otherwise returns the tilemap

Run the wave function collapse algorithm on a heightmap. The algorithm will generate a tilemap based on the heightmap and the rules of the tiles. The algorithm will return a tilemap that satisfies the rules of the tiles.