Utils

Documentation

listToShape :: [Pos] -> Shape

Convert a list of relative positions into a shape that takes a position and returns the absolute positions of the shape at that position

allNeighbours :: Shape

A shape that contains all the direct horizontal, vertical and diagonal neighbours of a position

nextToAny :: [Tile] -> Shape -> Rule

A rule that takes a list of tiles and a shape, and returns True if any of the tiles defined in the list tile are within the shape at the given position.

nextToAll :: [Tile] -> Shape -> Rule

A rule that takes a list of tiles and a shape, and returns True if all of the tiles defined in the list tile are within the shape at the given position.

chanceRule :: Float -> Rule

A rule that takes a float f and returns a rule with chance f of returning True

isInPos :: [Pos] -> Pos -> Bool

A function that takes a list of positions and returns a function that checks if a position is in the list

canExistAt :: (Pos -> Bool) -> Rule

A rule that takes a function that takes a position and returns a boolean, and returns a rule using this function